requirements.txt
# Polymarket API
py-clob-client>=0.34.1

# Crypto Exchange APIs
python-binance>=1.0.19
ccxt>=4.0.0

# Web3 and Blockchain
web3>=6.11.0
eth-account>=0.10.0

# WebSocket and Async
websockets>=12.0
aiohttp>=3.9.0
asyncio

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Machine Learning (for AI predictions)
scikit-learn>=1.3.0
tensorflow>=2.15.0
torch>=2.1.0

# API and Data
requests>=2.31.0
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Configuration
pydantic>=2.5.0
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
